:recycle: Refactor: AppContextHolder 폴더 변경
