+++ composeApp/src/androidMain/kotlin/com/dh/ondot/core/platform/KakaoSignIn.kt
@@ -2,7 +2,7 @@ package com.dh.ondot.core.platform
 
 import android.content.Context
 import android.util.Log
-import com.dh.ondot.util.AppContextHolder
+import com.dh.ondot.core.util.AppContextHolder
 import com.kakao.sdk.user.UserApiClient
 
 actual fun kakaoSignIn(onResult: (String) -> Unit) {
