+++ composeApp/src/androidMain/kotlin/com/dh/ondot/core/platform/PlatformDependencies.android.kt
@@ -26,7 +26,7 @@ import com.dh.ondot.domain.service.AlarmScheduler
 import com.dh.ondot.domain.service.AlarmStorage
 import com.dh.ondot.domain.service.MapProviderStorage
 import com.dh.ondot.domain.service.SoundPlayer
-import com.dh.ondot.util.AppContextHolder
+import com.dh.ondot.core.util.AppContextHolder
 
 actual fun provideTokenProvider(): TokenProvider {
     val context = runCatching { AppContextHolder.context }
